pyproject.toml
[project]
name = "mcp-api-gateway"
version = "0.1.0"
description = "Config-driven MCP to REST intelligent API gateway."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.21.0",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "jsonschema>=4.25.1",
    "pydantic>=2.11.7",
    "PyYAML>=6.0.2",
    "uvicorn[standard]>=0.35.0",
    "watchfiles>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
]

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
